bundle copc and other requested changes
bundles copc into one file
improve errornaming
